requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.6.3
attrs==25.4.0
bcrypt==4.3.0
beautifulsoup4==4.14.2
blivet==3.12.1
blivet-gui==2.6.0
boto3==1.40.72
botocore==1.40.72
Brlapi==0.8.7
Brotli==1.1.0
certifi==2025.7.9
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-351-build/cockpit-351/tmp/wheel/cockpit-351-py3-none-any.whl#sha256=59d2dcd42c86f8d88df2a0b4afd6bee5c58753efde42f78d6ed1686b14aa61df
contourpy==1.3.3
crypt_r==3.13.1
cryptography==45.0.4
cupshelpers==1.0
cycler==0.11.0
dasbus==1.7
dbus-python==1.4.0
dbus_next==0.2.3
distro==1.9.0
dnf==4.24.0
dnspython==2.8.0
email-validator==2.3.0
fedora-third-party==0.10
file-magic==0.4.0
fonttools==4.60.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
idna==3.10
initial-setup==0.3.101
iso639==0.1.4
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.9
langtable==0.0.69
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.17.0
louis==3.33.0
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.6
multidict==6.6.4
mutagen==1.47.0
nftables==0.1
numpy==2.3.4
olefile==0.47
ollama==0.4.7
packaging==25.0
perf==0.1
pexpect==4.9.0
pid==3.0.4
pillow==11.3.0
ply==3.11
productmd==1.48
progressbar2==4.5.0
propcache==0.4.1
proton-core==0.3.3
proton-vpn-api-core==0.36.6
proton-vpn-network-manager==0.10.2
psutil==7.0.0
ptyprocess==0.7.0
pwquality==1.4.5
PyAudio==0.2.13
pycairo==1.28.0
pycares==4.10.0
pycparser==2.22
pycryptodomex==3.23.0
pycups==2.0.4
pydantic==2.12.4
pydantic_core==2.41.5
pyenchant==3.2.2
pygame==2.6.1
pygdbmi==0.11.0.0
PyGObject==3.54.5
pyinotify==0.9.6
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
PyNaCl==1.5.0
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
pyOpenSSL==25.1.0
pyparsing==3.1.2
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-gnupg==0.5.4
python-linux-procfs==0.7.3
python-meh==0.52
python-utils==3.9.1
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.17.8-build/kernel-6.17.8/linux-6.17.8-300.fc43.x86_64/tools/net/ynl
RapidFuzz==3.12.2
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
rpds-py==0.27.0
rpm==6.0.0
rpmautospec==0.8.3
rpmautospec-core==0.1.5
s3transfer==0.14.0
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
sentry-sdk==2.35.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
setools==4.6.0
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.35-build/setroubleshoot-3.3.35/src
setuptools==78.1.1
shtab==1.7.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.10.1
soupsieve==2.8
systemd-python==235
termcolor==3.1.0
tldr==3.4.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
xkbregistry==0.3
yarl==1.22.0
yt-dlp==2025.10.22
